maintaining constant table names wherever required; moving transaction related functions to transaction file; returning array of Transactions struct instead of byte array
