+++ constants/tables.go
@@ -1,0 +1,12 @@
+package constants
+
+const (
+	TokenchainTable = "tokenchain"
+	TokenchainIndexTable = "tokenchain_index"
+)
+
+// fullnode table names
+const (
+	FullnodeTokenchainTable = "fullnode_tokenchain"
+	FullnodeTokenchainIndexTable = "fullnode_tokenchain_index"
+)
